fix(recurring): move calculateNextDate out of use server boundary
Turbopack rejected the synchronous export from a "use server" file.
Extracted calculateNextDate and addMonthsUTC to lib/recurring-utils.ts
(no server directive). recurring.ts imports and re-exports the function
so existing callers are unaffected. Removed erroneous await on the
now-explicit sync call inside generateRecurringInvoices.
